+++ src/constants/Extension.ts
@@ -18,7 +18,6 @@ export const COMMAND_NAME = {
   setLastModifiedDate: getCommandName('setLastModifiedDate'),
   generateSlug: getCommandName('generateSlug'),
   createFromTemplate: getCommandName('createFromTemplate'),
-  toggleDraft: getCommandName('toggleDraft'),
   registerFolder: getCommandName('registerFolder'),
   unregisterFolder: getCommandName('unregisterFolder'),
   createContent: getCommandName('createContent'),